+++ sql/emaj--3.2.0--devel.sql
@@ -98,6 +98,390 @@ SELECT emaj._disable_event_triggers();
 ------------------------------------------------------------------
 -- create new or modified functions                             --
 ------------------------------------------------------------------
+CREATE OR REPLACE FUNCTION emaj._create_seq(v_schema TEXT, v_seq TEXT, v_groupName TEXT, v_timeId BIGINT)
+RETURNS VOID LANGUAGE plpgsql AS
+$_create_seq$
+-- The function records a sequence into a tables group
+-- Required inputs: the application sequence to process, the group to add it into, the priority attribute, the time id of the operation.
+  BEGIN
+-- record the sequence in the emaj_relation table
+    INSERT INTO emaj.emaj_relation (rel_schema, rel_tblseq, rel_time_range, rel_group, rel_kind)
+      VALUES (v_schema, v_seq, int8range(v_timeId, NULL, '[)'), v_groupName, 'S');
+    RETURN;
+  END;
+$_create_seq$;
+
+CREATE OR REPLACE FUNCTION emaj._verify_all_groups()
+RETURNS SETOF TEXT LANGUAGE plpgsql AS
+$_verify_all_groups$
+-- The function verifies the consistency of all E-Maj groups.
+-- It returns a set of error or warning messages for discovered discrepancies.
+-- If no error is detected, no row is returned.
+  BEGIN
+--
+-- Errors detection
+--
+-- check the postgres version at groups creation time is compatible (i.e. >= 8.4)
+    RETURN QUERY
+      SELECT 'The group "' || group_name || '" has been created with a non compatible postgresql version (' ||
+               group_pg_version || '). It must be dropped and recreated.' AS msg
+        FROM emaj.emaj_group
+        WHERE cast(to_number(substring(group_pg_version FROM E'^(\\d+)'),'99') * 100 +
+                   to_number(substring(group_pg_version FROM E'^\\d+\\.(\\d+)'),'99') AS INTEGER) < 804
+        ORDER BY msg;
+-- check all application schemas referenced in the emaj_relation table still exist
+    RETURN QUERY
+      SELECT 'The application schema "' || rel_schema || '" does not exist any more.' AS msg
+        FROM (
+          SELECT DISTINCT rel_schema FROM emaj.emaj_relation
+            WHERE upper_inf(rel_time_range)
+            EXCEPT
+          SELECT nspname FROM pg_catalog.pg_namespace
+             ) AS t
+        ORDER BY msg;
+-- check all application relations referenced in the emaj_relation table still exist
+    RETURN QUERY
+      SELECT 'In group "' || r.rel_group || '", the ' ||
+               CASE WHEN t.rel_kind = 'r' THEN 'table "' ELSE 'sequence "' END ||
+               t.rel_schema || '"."' || t.rel_tblseq || '" does not exist any more.' AS msg
+        FROM (                                        -- all expected application relations
+          SELECT rel_schema, rel_tblseq, rel_kind FROM emaj.emaj_relation
+            WHERE upper_inf(rel_time_range)
+            EXCEPT                                    -- minus relations known by postgres
+          SELECT nspname, relname, relkind FROM pg_catalog.pg_class, pg_catalog.pg_namespace
+            WHERE relnamespace = pg_namespace.oid AND relkind IN ('r','S')
+             ) AS t, emaj.emaj_relation r             -- join with emaj_relation to get the group name
+        WHERE t.rel_schema = r.rel_schema AND t.rel_tblseq = r.rel_tblseq AND upper_inf(r.rel_time_range)
+        ORDER BY t.rel_schema, t.rel_tblseq, 1;
+-- check the log table for all tables referenced in the emaj_relation table still exist
+    RETURN QUERY
+      SELECT 'In group "' || rel_group || '", the log table "' ||
+               rel_log_schema || '"."' || rel_log_table || '" is not found.' AS msg
+        FROM emaj.emaj_relation
+        WHERE rel_kind = 'r'
+          AND NOT EXISTS
+              (SELECT NULL FROM pg_catalog.pg_namespace, pg_catalog.pg_class
+                 WHERE nspname = rel_log_schema AND relname = rel_log_table
+                   AND relnamespace = pg_namespace.oid)
+        ORDER BY rel_schema, rel_tblseq, 1;
+-- check the log sequence for all tables referenced in the emaj_relation table still exist
+    RETURN QUERY
+      SELECT 'In group "' || rel_group || '", the log sequence "' ||
+               rel_log_schema || '"."' || rel_log_sequence || '" is not found.' AS msg
+        FROM emaj.emaj_relation
+        WHERE upper_inf(rel_time_range) AND rel_kind = 'r'
+          AND NOT EXISTS
+              (SELECT NULL FROM pg_catalog.pg_namespace, pg_catalog.pg_class
+                 WHERE nspname = rel_log_schema AND relname = rel_log_sequence
+                   AND relnamespace = pg_namespace.oid)
+        ORDER BY rel_schema, rel_tblseq, 1;
+-- check the log function for each table referenced in the emaj_relation table still exist
+    RETURN QUERY
+      SELECT 'In group "' || rel_group || '", the log function "' || rel_log_schema || '"."' || rel_log_function || '" is not found.'
+             AS msg
+        FROM emaj.emaj_relation
+        WHERE upper_inf(rel_time_range) AND rel_kind = 'r'
+          AND NOT EXISTS
+              (SELECT NULL FROM pg_catalog.pg_proc, pg_catalog.pg_namespace
+                 WHERE nspname = rel_log_schema AND proname = rel_log_function
+                   AND pronamespace = pg_namespace.oid)
+        ORDER BY rel_schema, rel_tblseq, 1;
+-- check log and truncate triggers for all tables referenced in the emaj_relation table still exist
+--   start with log triggers
+    RETURN QUERY
+      SELECT 'In group "' || rel_group || '", the log trigger "emaj_log_trg" on table "' ||
+               rel_schema || '"."' || rel_tblseq || '" is not found.' AS msg
+        FROM emaj.emaj_relation
+        WHERE upper_inf(rel_time_range) AND rel_kind = 'r'
+          AND NOT EXISTS
+              (SELECT NULL FROM pg_catalog.pg_trigger, pg_catalog.pg_namespace, pg_catalog.pg_class
+                 WHERE nspname = rel_schema AND relname = rel_tblseq AND tgname = 'emaj_log_trg'
+                   AND tgrelid = pg_class.oid AND relnamespace = pg_namespace.oid)
+                       -- do not issue a row if the application table does not exist,
+                       -- this case has been already detected
+          AND EXISTS
+              (SELECT NULL FROM pg_catalog.pg_class, pg_catalog.pg_namespace
+                 WHERE nspname = rel_schema AND relname = rel_tblseq AND relnamespace = pg_namespace.oid)
+        ORDER BY rel_schema, rel_tblseq, 1;
+--   then truncate triggers
+    RETURN QUERY
+      SELECT 'In group "' || rel_group || '", the truncate trigger "emaj_trunc_trg" on table "' ||
+             rel_schema || '"."' || rel_tblseq || '" is not found.' AS msg
+        FROM emaj.emaj_relation
+        WHERE upper_inf(rel_time_range) AND rel_kind = 'r'
+          AND NOT EXISTS
+              (SELECT NULL FROM pg_catalog.pg_trigger, pg_catalog.pg_namespace, pg_catalog.pg_class
+                 WHERE nspname = rel_schema AND relname = rel_tblseq AND tgname = 'emaj_trunc_trg'
+                   AND tgrelid = pg_class.oid AND relnamespace = pg_namespace.oid)
+                       -- do not issue a row if the application table does not exist,
+                       -- this case has been already detected
+          AND EXISTS
+              (SELECT NULL FROM pg_catalog.pg_class, pg_catalog.pg_namespace
+                 WHERE nspname = rel_schema AND relname = rel_tblseq AND relnamespace = pg_namespace.oid)
+        ORDER BY rel_schema, rel_tblseq, 1;
+-- check all log tables have a structure consistent with the application tables they reference
+--      (same columns and same formats). It only returns one row per faulting table.
+    RETURN QUERY
+      SELECT msg FROM (
+        WITH cte_app_tables_columns AS (                -- application table's columns
+            SELECT rel_group, rel_schema, rel_tblseq, rel_log_schema, rel_log_table, attname, atttypid, attlen, atttypmod
+              FROM emaj.emaj_relation, pg_catalog.pg_attribute, pg_catalog.pg_class, pg_catalog.pg_namespace
+              WHERE relnamespace = pg_namespace.oid AND nspname = rel_schema AND relname = rel_tblseq
+                AND attrelid = pg_class.oid AND attnum > 0 AND attisdropped = FALSE
+                AND upper_inf(rel_time_range) AND rel_kind = 'r'),
+             cte_log_tables_columns AS (                -- log table's columns
+            SELECT rel_group, rel_schema, rel_tblseq, rel_log_schema, rel_log_table, attname, atttypid, attlen, atttypmod
+              FROM emaj.emaj_relation, pg_catalog.pg_attribute, pg_catalog.pg_class, pg_catalog.pg_namespace
+              WHERE relnamespace = pg_namespace.oid AND nspname = rel_log_schema
+                AND relname = rel_log_table
+                AND attrelid = pg_class.oid AND attnum > 0 AND attisdropped = FALSE AND attnum < rel_emaj_verb_attnum
+                AND upper_inf(rel_time_range) AND rel_kind = 'r')
+        SELECT DISTINCT rel_schema, rel_tblseq,
+               'In group "' || rel_group || '", the structure of the application table "' ||
+                 rel_schema || '"."' || rel_tblseq || '" is not coherent with its log table ("' ||
+               rel_log_schema || '"."' || rel_log_table || '").' AS msg
+          FROM (
+            (                                           -- application table's columns
+            SELECT rel_group, rel_schema, rel_tblseq, rel_log_schema, rel_log_table, attname, atttypid, attlen, atttypmod
+              FROM cte_app_tables_columns
+            EXCEPT                                      -- minus log table's columns
+            SELECT rel_group, rel_schema, rel_tblseq, rel_log_schema, rel_log_table, attname, atttypid, attlen, atttypmod
+              FROM cte_log_tables_columns
+            )
+            UNION
+            (                                           -- log table's columns
+            SELECT rel_group, rel_schema, rel_tblseq, rel_log_schema, rel_log_table, attname, atttypid, attlen, atttypmod
+              FROM cte_log_tables_columns
+            EXCEPT                                      --  minus application table's columns
+            SELECT rel_group, rel_schema, rel_tblseq, rel_log_schema, rel_log_table, attname, atttypid, attlen, atttypmod
+              FROM cte_app_tables_columns
+            )) AS t
+                           -- do not issue a row if the log or application table does not exist,
+                           -- these cases have been already detected
+        WHERE (rel_log_schema, rel_log_table) IN
+              (SELECT nspname, relname FROM pg_catalog.pg_class, pg_catalog.pg_namespace
+                 WHERE relnamespace = pg_namespace.oid)
+          AND (rel_schema, rel_tblseq) IN
+              (SELECT nspname, relname FROM pg_catalog.pg_class, pg_catalog.pg_namespace
+                 WHERE relnamespace = pg_namespace.oid)
+        ORDER BY 1,2,3
+        ) AS t;
+-- check all tables of rollbackable groups have their primary key
+    RETURN QUERY
+      SELECT 'In rollbackable group "' || rel_group || '", the table "' ||
+             rel_schema || '"."' || rel_tblseq || '" has no primary key any more.' AS msg
+        FROM emaj.emaj_relation, emaj.emaj_group
+        WHERE upper_inf(rel_time_range) AND rel_kind = 'r' AND rel_group = group_name AND group_is_rollbackable
+          AND NOT EXISTS
+              (SELECT NULL FROM pg_catalog.pg_class, pg_catalog.pg_namespace, pg_catalog.pg_constraint
+                 WHERE nspname = rel_schema AND relname = rel_tblseq
+                   AND relnamespace = pg_namespace.oid AND connamespace = pg_namespace.oid AND conrelid = pg_class.oid
+                   AND contype = 'p')
+                       -- do not issue a row if the application table does not exist,
+                       -- this case has been already detected
+          AND EXISTS
+              (SELECT NULL FROM pg_catalog.pg_class, pg_catalog.pg_namespace
+                 WHERE nspname = rel_schema AND relname = rel_tblseq AND relnamespace = pg_namespace.oid)
+        ORDER BY rel_schema, rel_tblseq, 1;
+-- for rollbackable groups, check no table has been altered as UNLOGGED or dropped and recreated as TEMP table after tables groups creation
+    RETURN QUERY
+      SELECT 'In rollbackable group "' || rel_group || '", the table "' ||
+             rel_schema || '"."' || rel_tblseq || '" is UNLOGGED or TEMP.' AS msg
+        FROM emaj.emaj_relation, pg_catalog.pg_class, pg_catalog.pg_namespace, emaj.emaj_group
+        WHERE upper_inf(rel_time_range) AND rel_kind = 'r'
+          AND relnamespace = pg_namespace.oid AND nspname = rel_schema AND relname = rel_tblseq
+          AND group_name = rel_group AND group_is_rollbackable
+          AND relpersistence <> 'p'
+        ORDER BY rel_schema, rel_tblseq, 1;
+-- with PG 11-, check all tables are WITHOUT OIDS (i.e. have not been altered as WITH OIDS after their tables group creation)
+    IF emaj._pg_version_num() < 120000 THEN
+      RETURN QUERY
+        SELECT 'In rollbackable group "' || rel_group || '", the table "' ||
+               rel_schema || '"."' || rel_tblseq || '" is WITH OIDS.' AS msg
+          FROM emaj.emaj_relation, pg_catalog.pg_class, pg_catalog.pg_namespace, emaj.emaj_group
+          WHERE upper_inf(rel_time_range) AND rel_kind = 'r'
+            AND relnamespace = pg_namespace.oid AND nspname = rel_schema AND relname = rel_tblseq
+            AND group_name = rel_group AND group_is_rollbackable
+            AND relhasoids
+          ORDER BY rel_schema, rel_tblseq, 1;
+    END IF;
+-- check the primary key structure of all tables belonging to rollbackable groups is unchanged
+    RETURN QUERY
+      SELECT 'In rollbackable group "' || rel_group || '", the primary key of the table "' ||
+             rel_schema || '"."' || rel_tblseq || '" has changed (' || rel_sql_pk_columns || ' => ' || current_pk_columns || ').' AS msg
+        FROM (
+          SELECT rel_schema, rel_tblseq, rel_group, rel_sql_pk_columns,
+                 string_agg(quote_ident(attname), ',' ORDER BY attnum) AS current_pk_columns
+            FROM emaj.emaj_relation, emaj.emaj_group, pg_catalog.pg_attribute, pg_catalog.pg_index, pg_catalog.pg_class,
+                 pg_catalog.pg_namespace
+            WHERE -- join conditions
+                  rel_group = group_name
+              AND relname = rel_tblseq AND nspname = rel_schema
+              AND pg_attribute.attrelid = pg_index.indrelid
+              AND indrelid = pg_class.oid AND relnamespace = pg_namespace.oid
+                  -- filter conditions
+              AND rel_kind = 'r' AND upper_inf(rel_time_range)
+              AND group_is_rollbackable
+              AND attnum = ANY (indkey)
+              AND indisprimary
+              AND attnum > 0 AND attisdropped = FALSE
+            GROUP BY rel_schema, rel_tblseq, rel_group, rel_sql_pk_columns
+          ) AS t
+          WHERE rel_sql_pk_columns <> current_pk_columns
+        ORDER BY rel_schema, rel_tblseq, 1;
+-- check all log tables have the 6 required technical columns.
+    RETURN QUERY
+      SELECT msg FROM (
+        SELECT DISTINCT rel_schema, rel_tblseq,
+               'In group "' || rel_group || '", the log table "' ||
+               rel_log_schema || '"."' || rel_log_table || '" miss some technical columns (' ||
+               string_agg(attname,', ') || ').' AS msg
+          FROM (
+              SELECT rel_group, rel_schema, rel_tblseq, rel_log_schema, rel_log_table, attname
+                FROM emaj.emaj_relation,
+                     (VALUES ('emaj_verb'), ('emaj_tuple'), ('emaj_gid'), ('emaj_changed'), ('emaj_txid'), ('emaj_user')) AS t(attname)
+                WHERE rel_kind = 'r' AND upper_inf(rel_time_range)
+                  AND EXISTS
+                    (SELECT NULL FROM pg_catalog.pg_namespace, pg_catalog.pg_class
+                       WHERE nspname = rel_log_schema AND relname = rel_log_table
+                         AND relnamespace = pg_namespace.oid)
+            EXCEPT
+              SELECT rel_group, rel_schema, rel_tblseq, rel_log_schema, rel_log_table, attname
+                FROM emaj.emaj_relation, pg_catalog.pg_attribute, pg_catalog.pg_class, pg_catalog.pg_namespace
+                WHERE relnamespace = pg_namespace.oid AND nspname = rel_log_schema
+                  AND relname = rel_log_table
+                  AND attrelid = pg_class.oid AND attnum > 0 AND attisdropped = FALSE
+                  AND attname IN ('emaj_verb', 'emaj_tuple', 'emaj_gid', 'emaj_changed', 'emaj_txid', 'emaj_user')
+                  AND rel_kind = 'r' AND upper_inf(rel_time_range)
+             ) AS t2
+          GROUP BY rel_group, rel_schema, rel_tblseq, rel_log_schema, rel_log_table
+          ORDER BY 1,2,3
+        ) AS t;
+--
+-- Warnings detection
+--
+-- check all sequences associated to a serial or a "generated as identity" column have their related table in the same group
+    RETURN QUERY
+      SELECT msg FROM (
+        WITH serial_dependencies AS (
+          SELECT rs.rel_group AS seq_group, rs.rel_schema AS seq_schema, rs.rel_tblseq AS seq_name,
+                 rt.rel_group AS tbl_group, nt.nspname AS tbl_schema, ct.relname AS tbl_name
+            FROM emaj.emaj_relation rs
+                 JOIN pg_catalog.pg_class cs ON cs.relname = rel_tblseq
+                 JOIN pg_catalog.pg_namespace ns ON cs.relnamespace = ns.oid AND ns.nspname = rel_schema
+                 JOIN pg_depend ON pg_depend.objid = cs.oid
+                 JOIN pg_catalog.pg_class ct ON pg_depend.refobjid = ct.oid
+                 JOIN pg_catalog.pg_namespace nt ON ct.relnamespace = nt.oid
+                 LEFT OUTER JOIN emaj.emaj_relation rt ON rt.rel_schema = nt.nspname AND rt.rel_tblseq = ct.relname
+            WHERE rs.rel_kind = 'S' AND upper_inf(rs.rel_time_range)
+              AND (rt.rel_time_range IS NULL OR upper_inf(rt.rel_time_range))
+              AND pg_depend.classid = pg_depend.refclassid             -- the classid et refclassid must be 'pg_class'
+              AND pg_depend.classid = (SELECT oid FROM pg_catalog.pg_class WHERE relname = 'pg_class')
+        )
+        SELECT DISTINCT seq_schema, seq_name,
+               'Warning: In group "' || seq_group || '", the sequence "' || seq_schema || '"."' || seq_name ||
+               '" is linked to the table "' || tbl_schema || '"."' || tbl_name ||
+               '" but this table does not belong to any tables group.' AS msg
+          FROM serial_dependencies
+          WHERE tbl_group IS NULL
+        UNION ALL
+        SELECT DISTINCT seq_schema, seq_name,
+               'Warning: In group "' || seq_group || '", the sequence "' || seq_schema || '"."' || seq_name ||
+               '" is linked to the table "' || tbl_schema || '"."' || tbl_name ||
+               '" but this table belongs to another tables group (' || tbl_group || ').' AS msg
+          FROM serial_dependencies
+          WHERE tbl_group <> seq_group
+        ORDER BY 1,2,3
+        ) AS t;
+--
+    RETURN;
+  END;
+$_verify_all_groups$;
+
+CREATE OR REPLACE FUNCTION emaj._adjust_group_properties()
+RETURNS INTEGER LANGUAGE plpgsql
+SECURITY DEFINER SET search_path = pg_catalog, pg_temp AS
+$_adjust_group_properties$
+-- The function adjusts the content of the emaj_group table.
+-- It actualy only adjusts the group_has_waiting_changes column.
+-- This group_has_waiting_changes column is normally set by a trigger on emaj_group_def.
+-- But in some cases, its value may not correspond to the real situation. This function sets its value to the proper value.
+-- It mainly joins the content of the emaj_group_def and the emaj_relation table to detect differences.
+-- It also calls the _verify_groups() function to detect potential corrupted groups that would need to be altered.
+-- If needed, the emaj_group table is updated.
+-- The function is declared SECURITY DEFINER so that emaj_viewer roles can execute it when calling the emaj_verify_all() function
+-- It returns the number of groups that have been updated.
+  DECLARE
+    v_nbAdjustedGroups       INT = 0;
+  BEGIN
+-- process the group_has_waiting_changes column using one big SQL statement
+    WITH
+      tblseq_with_changes AS (
+        -- tables and sequences modified or deleted from emaj_group_def
+        SELECT rel_group, rel_schema, rel_tblseq
+          FROM emaj.emaj_relation
+               LEFT OUTER JOIN emaj.emaj_group_def ON (rel_schema = grpdef_schema AND rel_tblseq = grpdef_tblseq
+                                                       AND rel_group = grpdef_group)
+          WHERE upper_inf(rel_time_range)
+            AND (
+              -- the relations that do not belong to the groups anymore
+                  grpdef_group IS NULL
+              -- the tables whose log data tablespace in emaj_group_def has changed
+              --         or whose log index tablespace in emaj_group_def has changed
+               OR (rel_kind = 'r'
+                  AND (coalesce(rel_log_dat_tsp,'') <> coalesce(grpdef_log_dat_tsp,'')
+                    OR coalesce(rel_log_idx_tsp,'') <> coalesce(grpdef_log_idx_tsp,'')
+                      ))
+              -- the tables or sequences that change their group ownership
+               OR (rel_group <> grpdef_group)
+              -- the relation that change their priority level
+               OR (rel_priority IS NULL AND grpdef_priority IS NOT NULL) OR
+                  (rel_priority IS NOT NULL AND grpdef_priority IS NULL) OR
+                  (rel_priority <> grpdef_priority)
+                )
+      UNION
+        -- new tables or sequences in emaj_group_def
+        SELECT grpdef_group, grpdef_schema, grpdef_tblseq
+          FROM emaj.emaj_group_def, pg_catalog.pg_class, pg_catalog.pg_namespace, emaj.emaj_group
+          WHERE NOT EXISTS (
+                SELECT NULL FROM emaj.emaj_relation
+                  WHERE rel_schema = grpdef_schema AND rel_tblseq = grpdef_tblseq AND upper_inf(rel_time_range))
+            AND relnamespace = pg_namespace.oid AND nspname = grpdef_schema AND relname = grpdef_tblseq
+            AND group_name = grpdef_group
+      UNION
+        -- dammaged tables
+        SELECT ver_group, ver_schema, ver_tblseq
+          FROM emaj._verify_groups(
+                 (SELECT array_agg(group_name) FROM emaj.emaj_group)
+                 , false)
+          WHERE ver_group IS NOT NULL
+      ),
+      -- get the list of groups that would need to be altered
+      group_with_changes AS (
+        SELECT DISTINCT rel_group AS group_name
+          FROM tblseq_with_changes
+      ),
+      -- adjust the group_has_waiting_changes column, only when needed
+      modified_group AS (
+        UPDATE emaj.emaj_group SET group_has_waiting_changes = NOT group_has_waiting_changes
+          WHERE (group_has_waiting_changes = FALSE
+                 AND group_name IN (SELECT group_name FROM group_with_changes))
+             OR (group_has_waiting_changes = TRUE
+                 AND NOT EXISTS (SELECT 0 FROM group_with_changes WHERE group_with_changes.group_name = emaj_group.group_name))
+          RETURNING group_name, group_has_waiting_changes
+      ),
+      -- insert a row in the history for each flag change
+      hist_insert AS (
+        INSERT INTO emaj.emaj_hist (hist_function, hist_object, hist_wording)
+          SELECT 'ADJUST_GROUP_PROPERTIES', group_name, 'Set the group_has_waiting_changes column to ' || group_has_waiting_changes
+            FROM modified_group
+            ORDER BY group_name
+      )
+      SELECT count(*) INTO v_nbAdjustedGroups FROM modified_group;
+    RETURN v_nbAdjustedGroups;
+  END;
+$_adjust_group_properties$;
+
 --<end_functions>                                pattern used by the tool that extracts and insert the functions definition
 ------------------------------------------
 --                                      --